log-area reverted, area to double
git-svn-id: file:///home/fules/svn/text_similarity/trunk@1252 ac99d0f3-1a01-e011-8a23-002586f6d316

+++ WordSetInternalNode.java
@@ -77,13 +77,13 @@ class WordSetInternalNode extends WordSetNode implements WordSet.Queueable {
 	public void add(String s) {
 		super.add(s);
 
-		float leastAreaIncrease = Float.POSITIVE_INFINITY;
-		float leastArea = Float.POSITIVE_INFINITY;
+		double leastAreaIncrease = Double.POSITIVE_INFINITY;
+		double leastArea = Double.POSITIVE_INFINITY;
 		WordSetNode bestFit = null;
 
 		for (Iterator<WordSetNode> it = children.iterator(); it.hasNext(); ) {
 			WordSetNode n = it.next();
-			float d = n.boundary.areaIncreaseIfAdded(s);
+			double d = n.boundary.areaIncreaseIfAdded(s);
 			if ((d < leastAreaIncrease) ||
 			    (( d == leastAreaIncrease) && (n.size() < leastArea))) {
 				leastAreaIncrease = d;
